fix(dualstack): logic error in IPv4 socket creation
This commit fixes a logic error in the socket creation for v6 and v4.
The error case is that an IPv6 only socket would prevent an additional
IPv4 socket to be created.

With this commit each variable it's made sure that at least a dualstack
IPv6 socket exists or an IPv4 and IPv6 socket.

Special case is macOS, while they report to support dualsocket, in fact
it doesn't work.

Move the logic to it's own function and add some test cases.

Fixes: https://github.com/rosenpass/rosenpass/issues/448

Signed-off-by: Paul Spooren <mail@aparcar.org>